refactor: move art prompts service
